Testing of BCF codec
-- Rev.d tribble
-- Minor code cleanup
-- BCF2 encoder / decoder use Double not Float internally everywhere
-- Generalized VC testing framework

+++ public/java/src/org/broadinstitute/sting/utils/codecs/bcf2/BCF2Decoder.java
@@ -199,8 +199,8 @@ public class BCF2Decoder {
         return BCF2Utils.readInt(bytesForEachInt, recordStream);
     }
 
-    public final float rawFloatToFloat(final int rawFloat) {
-        return Float.intBitsToFloat(rawFloat);
+    public final double rawFloatToFloat(final int rawFloat) {
+        return (double)Float.intBitsToFloat(rawFloat);
     }
 
     // ----------------------------------------------------------------------
